dashboard_app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
loguru==0.7.2
slowapi==0.1.9
pydantic-settings==2.3.0
polars==1.38.1
xlsx2csv==0.8.2
openpyxl==3.1.2
fastexcel==0.11.1
pyarrow==15.0.0
